requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
keras
python-dotenv
mysql-connector-python
pymysql
